+++ netzooe_eservice_api/__init__.py
@@ -1,1 +1,3 @@
+"""Netz OÖ eService API client."""
+
 from netzooe_eservice_api.version import __version__  # noqa: F401