+++ scripts/setup.sh
@@ -5,4 +5,4 @@ set -e
 cd "$(dirname "$0")/.."
 
 uv sync --locked --all-extras --dev
-uv run pre-commit install
+uv run prek install --force