+++ tests/test_api.py
@@ -1,20 +1,18 @@
 from typing import Any
-from typing import Final
 
 import pytest
 from aiohttp import ClientError
 from aioresponses import aioresponses
-from yarl import URL
+from syrupy import SnapshotAssertion
 
 from netzooe_eservice_api.api import NetzOOEeServiceAPI
 from netzooe_eservice_api.api import Pod
 from netzooe_eservice_api.constants import ConsentsStatus
+from netzooe_eservice_api.constants import ConsumptionsDimension
 from netzooe_eservice_api.constants import ConsumptionsProfilesBranch
 from netzooe_eservice_api.error import APIError
 from netzooe_eservice_api.error import AuthenticationError
 from netzooe_eservice_api.error import InvalidJsonError
-
-LOGIN_AND_LOGOUT_API_CALLS: Final[int] = 3
 
 
 async def create_client(mock_api: aioresponses, /, *, repeat: bool | int = False) -> NetzOOEeServiceAPI:
@@ -56,7 +54,7 @@ async def create_logged_in_client(mock_api: aioresponses, /, *, repeat: bool | i
 
 class TestHappyPathNetzOOEeServiceAPI:
     @pytest.mark.asyncio
-    async def test_login(self) -> None:
+    async def test_login(self, snapshot: SnapshotAssertion) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
 
@@ -64,41 +62,10 @@ class TestHappyPathNetzOOEeServiceAPI:
             await client.logout()
             assert client.xsrf_token == ""
 
-            login_key = ("POST", URL("https://eservice.netzooe.at/service/j_security_check"))
-            session_key = ("GET", URL("https://eservice.netzooe.at/service/v1.0/session"))
-
-            assert mock_api.requests is not None
-            assert login_key in mock_api.requests
-            assert session_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS
-            assert len(mock_api.requests[login_key]) == 1
-            assert len(mock_api.requests[session_key]) == 1
-
-            login_call = mock_api.requests[login_key][0]
-
-            assert login_call.kwargs["json"] == {
-                "j_username": "test",
-                "j_password": "test",
-            }
-
-            assert login_call.kwargs["headers"] == {
-                "User-Agent": "Mozilla/5.0",
-                "Accept": "application/json, text/plain, */*",
-                "client-id": "netzonline",
-                "Content-Type": "application/json",
-            }
-
-            session_call = mock_api.requests[session_key][0]
-
-            assert session_call.kwargs["headers"] == {
-                "User-Agent": "Mozilla/5.0",
-                "Accept": "application/json, text/plain, */*",
-                "client-id": "netzonline",
-            }
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
-    async def test_reconnect(self) -> None:
+    async def test_reconnect(self, snapshot: SnapshotAssertion) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api, repeat=2)
 
@@ -107,19 +74,12 @@ class TestHappyPathNetzOOEeServiceAPI:
 
             result = await client.dashboard()
             assert result == {}
-
             await client.logout()
 
-            dashboard_key = ("GET", URL("https://eservice.netzooe.at/service/v1.0/dashboard"))
-
-            assert mock_api.requests is not None
-            assert dashboard_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[dashboard_key]) == 2  # noqa: PLR2004
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
-    async def test_dashboard(self) -> None:
+    async def test_dashboard(self, snapshot: SnapshotAssertion) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
 
@@ -127,19 +87,12 @@ class TestHappyPathNetzOOEeServiceAPI:
 
             result = await client.dashboard()
             assert result == {}
-
             await client.logout()
 
-            dashboard_key = ("GET", URL("https://eservice.netzooe.at/service/v1.0/dashboard"))
-
-            assert mock_api.requests is not None
-            assert dashboard_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[dashboard_key]) == 1
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
-    async def test_dashboard_without_login(self) -> None:
+    async def test_dashboard_without_login(self, snapshot: SnapshotAssertion) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_client(mock_api)
 
@@ -147,16 +100,9 @@ class TestHappyPathNetzOOEeServiceAPI:
 
             result = await client.dashboard()
             assert result == {}
-
             await client.logout()
 
-            dashboard_key = ("GET", URL("https://eservice.netzooe.at/service/v1.0/dashboard"))
-
-            assert mock_api.requests is not None
-            assert dashboard_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[dashboard_key]) == 1
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize(
@@ -175,7 +121,9 @@ class TestHappyPathNetzOOEeServiceAPI:
             ),
         ],
     )
-    async def test_consents(self, status: list[ConsentsStatus] | ConsentsStatus | None, url: str) -> None:
+    async def test_consents(
+        self, snapshot: SnapshotAssertion, status: list[ConsentsStatus] | ConsentsStatus | None, url: str
+    ) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
 
@@ -183,16 +131,9 @@ class TestHappyPathNetzOOEeServiceAPI:
 
             result: list[dict[str, Any]] = await client.consents(status)
             assert result == []
-
             await client.logout()
 
-            consents_key = ("GET", URL(url))
-
-            assert mock_api.requests is not None
-            assert consents_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[consents_key]) == 1
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize(
@@ -209,7 +150,10 @@ class TestHappyPathNetzOOEeServiceAPI:
         ],
     )
     async def test_consumptions_profiles(
-        self, branch: list[ConsumptionsProfilesBranch] | ConsumptionsProfilesBranch | None, url: str
+        self,
+        snapshot: SnapshotAssertion,
+        branch: list[ConsumptionsProfilesBranch] | ConsumptionsProfilesBranch | None,
+        url: str,
     ) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
@@ -218,19 +162,12 @@ class TestHappyPathNetzOOEeServiceAPI:
 
             result: list[dict[str, Any]] = await client.consumptions_profiles(branch)
             assert result == []
-
             await client.logout()
 
-            consumptions_profiles_key = ("GET", URL(url))
-
-            assert mock_api.requests is not None
-            assert consumptions_profiles_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[consumptions_profiles_key]) == 1
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
-    async def test_contract_account(self) -> None:
+    async def test_contract_account(self, snapshot: SnapshotAssertion) -> None:
         with aioresponses() as mock_api:
             client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
 
@@ -238,28 +175,16 @@ class TestHappyPathNetzOOEeServiceAPI:
 
             result = await client.contract_accounts(business_partner_number="123", contract_account_number="345")
             assert result == {}
-
             await client.logout()
 
-            contract_account_key = ("GET", URL("https://eservice.netzooe.at/service/v1.0/contract-accounts/123/345"))
-
-            assert mock_api.requests is not None
-            assert contract_account_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[contract_account_key]) == 1
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
-    async def test_consumptions_profile(self) -> None:
-        with aioresponses() as mock_api:
-            client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
-
-            mock_api.post(
-                "https://eservice.netzooe.at/service/v1.0/consumptions/profile/active", status=200, payload=[]
-            )
-
-            result: list[dict[str, Any]] = await client.consumptions_profile(
-                pods=[
+    @pytest.mark.parametrize(
+        ("pods", "dimension"),
+        [
+            (
+                [
                     Pod(
                         contract_account_number="123",
                         energy_community_id="345",
@@ -269,41 +194,42 @@ class TestHappyPathNetzOOEeServiceAPI:
                         date_from="2026-01-01",
                         date_to="2026-01-30",
                     )
-                ]
-            )
-            assert result == []
+                ],
+                ConsumptionsDimension.ENERGY,
+            ),
+            (
+                [
+                    Pod(
+                        contract_account_number="123",
+                        profile_type="mocked-profile",
+                        best_available_granularity="mocked-granularity",
+                        meter_point_administration_number="AT123",
+                        date_from="2026-01-01",
+                        date_to="2026-01-30",
+                    )
+                ],
+                ConsumptionsDimension.ENERGY,
+            ),
+        ],
+    )
+    async def test_consumptions_profile(
+        self, snapshot: SnapshotAssertion, pods: list[Pod], dimension: ConsumptionsDimension
+    ) -> None:
+        with aioresponses() as mock_api:
+            client: NetzOOEeServiceAPI = await create_logged_in_client(mock_api)
 
+            mock_api.post(
+                "https://eservice.netzooe.at/service/v1.0/consumptions/profile/active", status=200, payload=[]
+            )
+
+            result: list[dict[str, Any]] = await client.consumptions_profile(pods=pods, dimension=dimension)
+            assert result == []
             await client.logout()
 
-            consumptions_profile_key = (
-                "POST",
-                URL("https://eservice.netzooe.at/service/v1.0/consumptions/profile/active"),
-            )
-
-            assert mock_api.requests is not None
-            assert consumptions_profile_key in mock_api.requests
-
-            assert len(mock_api.requests) == LOGIN_AND_LOGOUT_API_CALLS + 1
-            assert len(mock_api.requests[consumptions_profile_key]) == 1
-
-            consumptions_profile_call = mock_api.requests[consumptions_profile_key][0]
-
-            payload = consumptions_profile_call.kwargs["json"]
-
-            assert payload["dimension"] == "ENERGY"
-
-            pod: dict[str, Any] = payload["pods"][0]
-            assert pod["contractAccountNumber"] == "123"
-            assert pod["energyCommunityId"] == "345"
-            assert pod["type"] == "mocked-profile"
-            assert pod["bestAvailableGranularity"] == "mocked-granularity"
-            assert pod["meterPointAdministrationNumber"] == "AT123"
-            assert pod["timerange"]["from"] == "2026-01-01"
-            assert pod["timerange"]["to"] == "2026-01-30"
+            assert mock_api.requests == snapshot
 
 
 class TestUnhappyPathNetzOOEeServiceAPI:
-
     @pytest.mark.asyncio
     @pytest.mark.parametrize(
         ("exception", "status", "message"),
@@ -320,7 +246,9 @@ class TestUnhappyPathNetzOOEeServiceAPI:
             ),
         ],
     )
-    async def test_login_failed(self, exception: type[APIError], status: int, message: str) -> None:
+    async def test_login_failed(
+        self, snapshot: SnapshotAssertion, exception: type[APIError], status: int, message: str
+    ) -> None:
         with aioresponses() as mock_api:
             mock_api.post(
                 "https://eservice.netzooe.at/service/j_security_check",
@@ -338,29 +266,10 @@ class TestUnhappyPathNetzOOEeServiceAPI:
             assert str(error.value) == message
             assert client.xsrf_token == ""
 
-            login_key = ("POST", URL("https://eservice.netzooe.at/service/j_security_check"))
-
-            assert mock_api.requests is not None
-            assert len(mock_api.requests) == 1
-            assert login_key in mock_api.requests
-            assert len(mock_api.requests[login_key]) == 1
-
-            login_call = mock_api.requests[login_key][0]
-
-            assert login_call.kwargs["json"] == {
-                "j_username": "test",
-                "j_password": "test",
-            }
-
-            assert login_call.kwargs["headers"] == {
-                "User-Agent": "Mozilla/5.0",
-                "Accept": "application/json, text/plain, */*",
-                "client-id": "netzonline",
-                "Content-Type": "application/json",
-            }
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
-    async def test_session_failed(self) -> None:
+    async def test_session_failed(self, snapshot: SnapshotAssertion) -> None:
         with aioresponses() as mock_api:
             mock_api.post(
                 "https://eservice.netzooe.at/service/j_security_check",
@@ -381,19 +290,9 @@ class TestUnhappyPathNetzOOEeServiceAPI:
                 await client.login()
 
             assert str(error.value) == "401 Unauthorized: No permission -- see authorization schemes"
-
             assert client.xsrf_token == ""
 
-            login_key = ("POST", URL("https://eservice.netzooe.at/service/j_security_check"))
-            session_key = ("GET", URL("https://eservice.netzooe.at/service/v1.0/session"))
-
-            assert mock_api.requests is not None
-            assert login_key in mock_api.requests
-            assert session_key in mock_api.requests
-
-            assert len(mock_api.requests) == 2  # noqa: PLR2004
-            assert len(mock_api.requests[login_key]) == 1
-            assert len(mock_api.requests[session_key]) == 1
+            assert mock_api.requests == snapshot
 
     @pytest.mark.asyncio
     async def test_reconnect_failed(self) -> None: